离线管理、修改计划和测试计划

+++ src/features/vocabulary/redux/saga/planSaga.js
@@ -5,7 +5,9 @@ import * as Constant from "../../common/constant";
 import { store } from "../../../../redux/store";
 import { LOAD_TASKS_SUCCEED, LOAD_TASKS_FAIL, LOAD_TASKS_START } from '../action/homeAction';
 import {
-    CHANGE_VOCA_BOOK_SUCCEED, CHANGE_VOCA_BOOK_FAIL, LOAD_VOCA_BOOKS, CHANGE_VOCA_BOOK, CHANGE_VOCA_BOOK_START, LOAD_VOCA_BOOKS_FAIL, LOAD_VOCA_BOOKS_SUCCEED, LOAD_VOCA_BOOKS_START, MODIFY_LAST_LEARN_DATE,
+    CHANGE_VOCA_BOOK, CHANGE_VOCA_BOOK_START,
+    CHANGE_VOCA_BOOK_SUCCEED, CHANGE_VOCA_BOOK_FAIL,
+    MODIFY_LAST_LEARN_DATE, MODIFY_PLAN, MODIFY_PLAN_START, MODIFY_PLAN_SUCCEED, MODIFY_PLAN_FAIL,
 } from '../action/planAction';
 import { CLEAR_PLAY } from '../action/vocaPlayAction';
 import VocaTaskService from '../../service/VocaTaskService';
@@ -19,35 +21,47 @@ export function* createPlan(action) {
     yield put({ type: LOAD_TASKS_START })
     try {
         const res = yield Http.post("/plan/create", action.payload.plan)
-        const { plan, words, articles } = res.data
-        //清空先前数据，存储新数据到realm
-        const vtd = VocaTaskDao.getInstance()
-        const artDao = ArticleDao.getInstance()
-        vtd.deleteAll()
-        artDao.deleteAllArticles()
-        vtd.saveBookWords(words)
-        artDao.saveArticles(articles)
-        //清空未同步任务
-        Storage.clearMapForKey('notSyncTasks')
+        if (res.status === 200) {
+            const { plan, words, articles } = res.data
+            //清空先前数据，存储新数据到realm
+            const vtd = VocaTaskDao.getInstance()
+            const artDao = ArticleDao.getInstance()
+            vtd.deleteAll()
+            artDao.deleteAllArticles()
+            vtd.saveBookWords(words)
+            artDao.saveArticles(articles)
+            //清空未同步任务
+            Storage.clearMapForKey('notSyncTasks')
 
+            const vts = new VocaTaskService()
+            //加载今日数据 
+            const tasks = vts.getTodayTasks(null, plan.taskCount, plan.taskWordCount)
+            //构建文章任务
+            const articleTasks = VocaUtil.genArticleTasksByVocaTasks(
+                tasks.filter((item, i) => item.status === Constant.STATUS_0)
+            )
+            // 计算剩余天数
+            const leftDays = vts.countLeftDays(plan.taskCount, plan.taskWordCount)
 
-        //加载今日数据 
-        const tasks = new VocaTaskService().getTodayTasks(null, plan.taskCount, plan.taskWordCount)
-        //构建文章任务
-        const articleTasks = VocaUtil.genArticleTasksByVocaTasks(tasks)
-
-        if (store.getState().vocaPlay.task.normalType === Constant.BY_REAL_TASK) {
-            yield put({ type: CLEAR_PLAY })
-        }
-        yield put({ type: LOAD_TASKS_SUCCEED, payload: { tasks: tasks.concat(articleTasks) } })
-        yield put({
-            type: CHANGE_VOCA_BOOK_SUCCEED,
-            payload: {
-                plan
+            if (store.getState().vocaPlay.task.normalType === Constant.BY_REAL_TASK) {
+                yield put({ type: CLEAR_PLAY })
             }
-        })
-        //加载成功后，修改lastLearnDate
-        yield put({ type: MODIFY_LAST_LEARN_DATE, payload: { lastLearnDate: _util.getDayTime(0) } })
+            yield put({ type: LOAD_TASKS_SUCCEED, payload: { tasks: tasks.concat(articleTasks) } })
+            yield put({
+                type: CHANGE_VOCA_BOOK_SUCCEED,
+                payload: {
+                    plan,
+                    leftDays
+                }
+            })
+            //加载成功后，修改lastLearnDate
+            yield put({ type: MODIFY_LAST_LEARN_DATE, payload: { lastLearnDate: _util.getDayTime(0) } })
+        } else {
+            yield put({ type: CHANGE_VOCA_BOOK_FAIL })
+            yield put({ type: LOAD_TASKS_FAIL })
+        }
+
+
     } catch (err) {
         console.log(err)
         yield put({ type: CHANGE_VOCA_BOOK_FAIL })
@@ -55,10 +69,34 @@ export function* createPlan(action) {
     }
 }
 
+export function* modifyPlan(action) {
+    yield put({ type: MODIFY_PLAN_START })
+    try {
+        const res = yield Http.post("/plan/modify", action.payload.plan)
+        if (res.status === 200) {
+            yield put({
+                type: MODIFY_PLAN_SUCCEED,
+                payload: {
+                    plan: res.data,
+                    leftDays: action.payload.leftDays
+                }
+            })
+        } else {
+            yield put({ type: MODIFY_PLAN_FAIL })
+        }
+    } catch (err) {
+        console.log(err)
+        yield put({ type: MODIFY_PLAN_FAIL })
+    }
+
+}
 
 
 
-
-export function* watchPlan() {
+export function* watchCreatePlan() {
     yield takeLatest(CHANGE_VOCA_BOOK, createPlan)
+}
+
+export function* watchModifyPlan() {
+    yield takeLatest(MODIFY_PLAN, modifyPlan)
 }
